% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type1.R
\name{Create_date}
\alias{Create_date}
\title{Create_date}
\usage{
Create_date(YEAR, MONTH, DAY)
}
\arguments{
\item{YEAR}{Integer and represents the corresponding year.}

\item{MONTH}{Integer between 1 and 12. MONTH stands for the month in date.}

\item{DAY}{Integer between 1 and 31. DAY stands for the number of day and must be respect some contrainsts according to the month.}
}
\value{
Date object
}
\description{
This function takes three integers as year, month, day in inputs and creates date by blending together in date format "year-month-day" if the given year, month, day are correct.
}
\examples{
{
year  <-  c("1984", "2004", "2000")
month <-  c("11", "06", NA)
day   <- c("15", "05", NA)
my_date <- Create_date(year, month, day)
}
}
