% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type1.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(
  datalist,
  YEAR = "YEAR",
  MONTH = "MONTH",
  DAY = "DAY",
  LATITUDE = "LATITUDE",
  LONGITUDE = "LONGITUDE"
)
}
\arguments{
\item{datalist}{stands for data.frame whichcontains earthquake information.}

\item{YEAR}{Interger}

\item{MONTH}{INteger between 1 to 12}

\item{DAY}{Integer  between 1 to 31}

\item{LATITUDE}{Double}

\item{LONGITUDE}{Double}
}
\value{
data.frame object
}
\description{
eq_clean_data makes earthquake data in the required format. In which each column has its data in the correct format.
}
\examples{
{
my_data <- data.frame(
YEAR = c("1984", "2000", "2014"),
MONTH = c("1", "2", "12"),
DAY = c("19", "20", "24"),
LONGITUDE = c("12.344", "14.89", "13.345"),
LATITUDE = c("-8.00", "-8.89", "-9.345")
)
my_data_clean <- eq_clean_data(my_data)
}
}
\seealso{
Create_date
}
