% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type2.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(my_data)
}
\arguments{
\item{my_data}{stands for data.frame in which the columns contains information in order to annatate a map.
this my_data must have CLEAN_LOCATION_NAME, EQ_PRIMARY, DEATHS as columns names.}
}
\value{
vector of character
}
\description{
Create formated html label for annoting map from leaflet object
Annotation focuses and combines information on location, magnitude accordin richter scale
and the nomber of deaths recorded during the earthquake.
}
\examples{
{
my_data <- data.frame(
DEATHS = c(125, 12),
EQ_PRIMARY  = c(7.981792, 8.019300),
CLEAN_LOCATION_NAME = c("Bamako Coura", "Badialan")
)

eq_create_label(my_data)
}
}
