% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type1.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(datalist, LOCATION_NAME = "LOCATION_NAME")
}
\arguments{
\item{datalist}{data.frame}

\item{LOCATION_NAME}{column of datalist which contains the complete location name of each earthquake.}
}
\value{
data.frame
}
\description{
create CLEAN_LOCATION_NAME by Removing the country names from LOCATION_NAME
}
\examples{
{
datalist <- data.frame (LOCATION_NAME = c("Mali : bamako",
                                          "Guinee : kankan",
                                          "Mali : sikasso wassoulou" ))
datalist <- eq_location_clean(datalist)
}
}
\seealso{
remove_Country_name
}
