% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type2.R
\name{eq_map}
\alias{eq_map}
\title{eq_map allows to annotate on a map or leaflet object position and gives information through a htlm popup}
\usage{
eq_map(my_data, annot_col)
}
\arguments{
\item{my_data}{stands for data.frame or list which contains at least two columns denoted by LONGITUDE and LATITUDE}

\item{annot_col}{defines the column which will appear on the map in popup}
}
\value{
a leaflet or htmlwidget object which is an interactive map if \code{annot_col} is a column name of
\code{my_data}
}
\description{
eq_map allows to annotate on a map or leaflet object position and gives information through a htlm popup
}
\examples{
{
mydata <- data.frame(
LONGITUDE = c(12.585080, 12.651165),
LATITUDE = c(-7.981792, -8.019300),
Annotation = c("Bamako Coura", "Badialan")
)
mymap <- eq_map(mydata, "Annotation")
mymap
}
}
