% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type3.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom timeline is used to create a geom point in date time in order to localize an event on time.}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options: If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{show.legend}{ogical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().}

\item{...}{Other arguments passed on to layer().}
}
\value{
Geom or ggproto object to build
}
\description{
geom timeline is used to create a geom point in date time in order to localize an event on time.
}
\examples{
\dontrun{
ggplot(data = dataCapstone_clean, mapping = aes(x = DATE)) +
geom_timeline( mapping = aes(size  = EQ_PRIMARY)) +
}
}
