% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type3.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label adds text directly to the plot}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  n_max = 5,
  size = 1,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default)}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{show.legend}{if TRUE a legend is shown on the figure}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{n_max}{Integer and stands for the maximum number of event to show by country}

\item{size}{double and stands for the size of text}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{Other arguments passed on to layer()}
}
\value{
Geom text
}
\description{
geom_timeline_label adds text directly to the plot
}
\examples{
\dontrun{
ggplot(data = dataCapstone_clean, mapping = aes(x = DATE)) +
geom_timeline( mapping = aes(size  = EQ_PRIMARY)) +
 geom_timeline_label(data = dataCapstone_clean,
                     mapping = aes( x = DATE,
                                    color = TOTAL_DEATHS,
                                    label = CLEAN_LOCATION_NAME
                     )
 ) +
 ylim(0, 5)

}
}
