% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type3.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Defines Our GeomTimeline class}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
The ggproto() function is used to construct a new class corresponding to your new geom.
}
\examples{
\dontrun{
ggplot(dataCapstone_clean, aes(x = date, y = COUNTRY,
               color = as.numeric(TOTAL_DEATHS),
               size = as.numeric(EQ_PRIMARY),
               label = CLEAN_LOCATION_NAME)) +
  geom_timeline() +
  ggplot2::theme(panel.background = ggplot2::element_blank(),
        legend.position = "bottom",
        axis.title.y = ggplot2::element_blank()) + ggplot2::xlab("DATE")

}
}
\keyword{datasets}
