% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_type1.R
\name{remove_Country_name}
\alias{remove_Country_name}
\title{remove_Country_name}
\usage{
remove_Country_name(string)
}
\arguments{
\item{string}{character with this format "something: name of city"}
}
\value{
character
}
\description{
remove_Country_name extracts the city name from a text
}
\examples{
{
textcity <- c("Mali:bamako", "Cote ivoire : man")
citytext <- remove_Country_name(textcity)
}
}
